backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
scikit-image
scikit-fuzzy
numpy
matplotlib
pillow
pydantic-settings
python-dotenv
tensorflow-macos; platform_system == "Darwin" and platform_machine == "arm64"
tensorflow; platform_system != "Darwin" or platform_machine != "arm64"
